First commit
Added a threshold function in the proofs.api to send threshold.
Still needs to set a maximum value for the threshold check.
Needs to be tested on a dummy transaction until the real data is taken into account.
Made the necessary changes to all the files in order to test it the api.